move v2 into main module
