+++ torcharc/specs/merge/stack.yaml
